Implented 1:1 Relationship

+++ src/main/java/com/srilankagem/gembackend/gem/dto/GemStoneRequest.java
@@ -6,6 +6,8 @@ import com.srilankagem.gembackend.gem.models.GemType;
 import jakarta.validation.constraints.*;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Setter
 @Getter
 @NoArgsConstructor
@@ -25,7 +27,7 @@ public class GemStoneRequest {
 
     @NotNull
     @DecimalMin(value = "0.01", message = "Value must be greater than 0.01")
-    private Double caratWeight;
+    private BigDecimal caratWeight;
 
     @NotNull(message = "origin is required")
     private GemOrigin origin;
@@ -35,7 +37,7 @@ public class GemStoneRequest {
 
     @NotNull
     @DecimalMin(value = "0.01", message = "Value has to be greater than 0.01")
-    private Double pricePerCarat;
+    private BigDecimal pricePerCarat;
 
     @NotNull
     @Min(value = 0, message = "Value has to be greater than 0")
